web_interface/requirements.txt
Flask==2.3.3
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2
albumentations==1.3.1
Werkzeug==2.3.7 